+++ metro.config.js
@@ -1,23 +1,22 @@
+const { getDefaultConfig, mergeConfig } = require('@react-native/metro-config')
+
+const defaultConfig = getDefaultConfig(__dirname)
+const { assetExts, sourceExts } = defaultConfig.resolver
+
 /**
- * Metro configuration for React Native
- * https://github.com/facebook/react-native
+ * Metro configuration
+ * https://facebook.github.io/metro/docs/configuration
  *
- * @format
+ * @type {import('metro-config').MetroConfig}
  */
-
-const { getDefaultConfig } = require('metro-config')
-
-module.exports = (async () => {
-  const {
-    resolver: { sourceExts, assetExts }
-  } = await getDefaultConfig()
-  return {
-    transformer: {
-      babelTransformerPath: require.resolve('react-native-svg-transformer')
-    },
-    resolver: {
-      assetExts: assetExts.filter(ext => ext !== 'svg'),
-      sourceExts: [...sourceExts, 'svg']
-    }
+const config = {
+  transformer: {
+    babelTransformerPath: require.resolve('react-native-svg-transformer')
+  },
+  resolver: {
+    assetExts: assetExts.filter(ext => ext !== 'svg'),
+    sourceExts: [...sourceExts, 'svg']
   }
-})()
+}
+
+module.exports = mergeConfig(getDefaultConfig(__dirname), config)